+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.UUID;
 import kitchenpos.menu.domain.model.MenuGroup;
 import kitchenpos.menu.domain.model.MenuGroupName;
+import kitchenpos.menu.domain.model.MenuGroupNameCreationService;
 import kitchenpos.menu.domain.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,15 +12,18 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class MenuGroupService {
     private final MenuGroupRepository menuGroupRepository;
+    private final MenuGroupNameCreationService menuGroupNameCreationService;
 
-    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository, MenuGroupNameCreationService menuGroupNameCreationService) {
         this.menuGroupRepository = menuGroupRepository;
+        this.menuGroupNameCreationService = menuGroupNameCreationService;
     }
 
     @Transactional
     public MenuGroup create(final MenuGroup request) {
         final String name = request.getName();
-        final MenuGroup menuGroup = new MenuGroup(new MenuGroupName(name), UUID.randomUUID());
+        MenuGroupName menuGroupName = menuGroupNameCreationService.createName(name);
+        final MenuGroup menuGroup = new MenuGroup(menuGroupName, UUID.randomUUID());
         return menuGroupRepository.save(menuGroup);
     }
