+++ src/main/java/kitchenpos/menu/domain/model/MenuGroupNameCreationService.java
@@ -1,7 +1,6 @@
-package kitchenpos.menu.domain.service;
+package kitchenpos.menu.domain.model;
 
 import kitchenpos.common.application.PurgomalumClient;
-import kitchenpos.menu.domain.model.MenuGroupName;
 import org.springframework.stereotype.Service;
 
 @Service
